requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
PyMuPDF==1.24.9
Pillow==10.4.0
pydantic==2.8.2
pydantic-settings==2.3.4
aiofiles==24.1.0
pytest==8.2.2
httpx==0.27.0
